requirements.txt
# Paquetes base
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2

# Visualización
matplotlib==3.8.3
seaborn==0.13.2
pillow==10.2.0

# Procesamiento
scipy==1.13.0
imbalanced-learn==0.11.0
joblib==1.3.2

# Utilidades
tqdm==4.66.2
requests==2.31.0
urllib3==2.2.1

# PyTorch y dependencias (ya instaladas)
filelock==3.13.1
typing-extensions==4.9.0
Jinja2==3.1.3
MarkupSafe==2.1.5